adição de alguns testes para poligonos e subclasses de poligonos; remoção da pasta "TEST"

+++ src/Ponto.java
@@ -83,7 +83,7 @@ public class Ponto {
     @Override
     public String toString()
     {
-        return String.format("(%d,%d)", x, y);
+        return String.format("(%d,%d)", (int)x, (int)y);
     }
